Early draft for singularity/lmod, python reqs
(1) Added two guides that are very early writeups on how to get started
with lmod and singularity.
(2) Optimized the WSL2 install script for singularity so that checkouts
only happen when required.
(3) Added a split requirements.txt and requirements-dev.txt file for
the new earthquake code.  This will allow us to easily identify what
python libraries are needed for runtime, and what are good to have
installed for development.